+++ gametrail/api/Game/views.py
@@ -1,0 +1,115 @@
+from django.shortcuts import get_object_or_404
+from rest_framework.viewsets import ModelViewSet
+from rest_framework.permissions import IsAuthenticated
+from django.http import HttpResponse
+from gametrail import functions
+from django.db.models import Avg
+from gametrail.models import *
+from gametrail.api.serializers import *
+from gametrail.api.Game.gameSerializers import *
+from gametrail.api.User.userSerializers import *
+from gametrail.api.Trail.trailSerializers import *
+from gametrail.api.User.views import *
+from gametrail.api.Trail.views import *
+from django_filters.rest_framework import DjangoFilterBackend
+from rest_framework.filters import SearchFilter
+from rest_framework.decorators import action
+from rest_framework.response import Response
+from rest_framework import status, viewsets
+from rest_framework.views import APIView
+from rest_framework.authentication import TokenAuthentication
+from rest_framework.authtoken.views import ObtainAuthToken
+from rest_framework.authtoken.models import Token
+from rest_framework.response import Response
+from itertools import chain
+from django.db.models.query import QuerySet
+from datetime import datetime
+from django.core import serializers
+from rest_framework.pagination import PageNumberPagination
+
+def check_user_is_admin(request):
+    user = request.user
+    return user.is_staff
+def check_user_is_the_same(request,usergametrail):
+    user = request.user
+    return user.username == usergametrail.username
+
+def check_user_is_authenticated(request):
+    user = request.user
+    return user.is_authenticated
+class StandardResultsSetPagination(PageNumberPagination):
+    page_size = 16
+    page_size_query_param = 'page_size'
+    max_page_size = 16
+
+class GetGameApiViewSet(ModelViewSet):
+    http_method_names = ['get']
+    serializer_class = GetGameSerializer
+    queryset = Game.objects.all()
+    filter_backends = [SearchFilter, DjangoFilterBackend]
+    search_fields = ['name']
+    filterset_fields = ['platforms__platform','genres__genre']
+    pagination_class = StandardResultsSetPagination
+
+
+class GetRecentGames(ModelViewSet):
+    http_method_names = ['get']
+    serializer_class = GetGameSerializer
+    queryset = Game.objects.all().order_by("-id")[:10]
+    
+class CUDGameApiViewSet(APIView):
+    http_method_names = ['post', 'put', 'delete']
+    serializer_class = CUDGameSerializer
+
+    def post(self, request, format = None):
+        is_user_admin = check_user_is_admin(request)
+
+        if is_user_admin == False:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+        else:
+            serializer = CUDGameSerializer(data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data, status=status.HTTP_201_CREATED)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    def put(self, request, format = None):
+        is_user_admin = check_user_is_admin(request)
+
+        if is_user_admin == False:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+        else:
+            try:
+                game = Game.objects.get(pk=request.data['id'])
+            except Game.DoesNotExist:
+                return Response(status=status.HTTP_404_NOT_FOUND)
+            
+            serializer = CUDGameSerializer(game, data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data, status=status.HTTP_201_CREATED)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    def delete(self, request, format = None):
+        is_user_admin = check_user_is_admin(request)
+        if is_user_admin == False:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
+        else:
+            try:
+                game = Game.objects.get(pk=request.data['id'])
+            except Game.DoesNotExist:
+                return Response(status=status.HTTP_404_NOT_FOUND)
+            
+            game.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+
+
+class GameCommentAPIView(ModelViewSet):
+    http_method_names = ['get']
+    serializer_class = CommentsOfAGameSerializer
+
+    def get_queryset(self):
+        game_id = self.request.query_params.get('game_id', None)
+        queryset = Comment.objects.filter(game_id=game_id)
+        return queryset
+
