+++ gametrail/api/routers.py
@@ -1,5 +1,8 @@
 from rest_framework.routers import DefaultRouter
 from gametrail.api.views import *
+from gametrail.api.Game.views import *
+from gametrail.api.Trail.views import *
+from gametrail.api.User.views import *
 
 router_api = DefaultRouter()
 router_api.register(prefix='game', viewset=GetGameApiViewSet, basename='game')
