+++ src/urls.py
@@ -1,10 +1,12 @@
 from django.contrib import admin
 from django.urls import path, include
-from gametrail.api.views import Logout, CustomAuthToken, CUGameInListApiViewSet
-from gametrail.api.views import CUDGameApiViewSet, GetGameApiViewSet
+from gametrail.api.User.views import *
+from gametrail.api.Game.views import *
+from gametrail.api.Trail.views import *
 # Importar rutas de la API de prueba 'demoapi'
 from gametrail.api.routers import router_api
 from gametrail.api import views
+
 from rest_framework.authtoken import views as token_views
 
 urlpatterns = [
@@ -15,11 +17,11 @@ urlpatterns = [
     path('api/auth/login', CustomAuthToken.as_view(), name='token_obtain_pair'),
     path('api/game', CUDGameApiViewSet.as_view()),
     path('api/populate_sabias_que', views.populate_sabias_que),
-    path('api/comment', views.CUDCommentsAPIViewSet.as_view()),
-    path('api/rating', views.POSTRatingAPIViewSet.as_view()),
-    path('api/trail/', views.TrailApiViewSet.as_view()),
-    path('api/gameInTrail', views.GameInTrailViewSet.as_view()),
-    path('api/createMinRating', views.CreateMinRatingViewSet.as_view()),
-    path('api/addUserInTrail', views.AddUserInTrailViewSet.as_view()),
+    path('api/comment', CUDCommentsAPIViewSet.as_view()),
+    path('api/rating', POSTRatingAPIViewSet.as_view()),
+    path('api/trail/', TrailApiViewSet.as_view()),
+    path('api/gameInTrail', GameInTrailViewSet.as_view()),
+    path('api/createMinRating', CreateMinRatingViewSet.as_view()),
+    path('api/addUserInTrail', AddUserInTrailViewSet.as_view()),
     # path('api/trailRecomendation', views.UserTrailRecomendationViewSet.as_view())
 ]
